kafka-consumer-producer/requirements.txt
kafka-python
avro
avro-python3==1.10.2